Fix JavaDoc

+++ jhdf/src/main/java/io/jhdf/api/Dataset.java
@@ -46,7 +46,7 @@ public interface Dataset extends Node {
 	 *
 	 * @return the dimensions of this dataset
 	 */
-	int[] getDimensions();
+	long[] getDimensions();
 
 	/**
 	 * Checks if this dataset is scalar i.e is a single element with no dimensions.
@@ -81,8 +81,6 @@ public interface Dataset extends Node {
 	/**
 	 * Gets the max size of this dataset. If not specified this will be equal to
 	 * {@link #getDimensions()}
-	 *
-	 * If max size overflows an int ArithmeticException is thrown
 	 *
 	 * @return the max size of this dataset
 	 */
